0.6.2 🍦
added node and jest unit testing setup
