+++ .gitignore
@@ -1,3 +1,10 @@
+# Node
+node_modules
+
+# VS Code
+.vscode
+launch.json
+
 # Windows image file caches
 Thumbs.db
 ehthumbs.db
