+++ index.html
@@ -40,14 +40,13 @@
 
 </div>
 
-<!-- Global Definitions -->
-<script src="src/js/alphabet.js"></script>
-<script src="src/js/intervalLibrary.js"></script>
-
 <!-- Angular Code -->
 <!-- <script src="src/js/ng-index.js"></script> -->
 
-<!-- Methods -->
+<!-- Global Definitions -->
+<script src="src/js/global.js"></script>
+
+<!-- Methods Definitions -->
 <script src="src/js/getAlphabetFromRoot.js"></script>
 <script src="src/js/getUserIntervals.js "></script>
 <script src="src/js/getInterval.js "></script>
@@ -55,7 +54,6 @@
 <script src="src/js/getNoteChord.js"></script>
 <script src="src/js/updateChord.js"></script>
 <script src="src/js/getChord.js"></script>
-
 
 <script src="src/js/index.js"></script>
