+++ src/js/global.js
@@ -1,3 +1,35 @@
+
+//defining different names for notes even though the final web app won't have the ability to distinguish between them because the user will only have a keyboard layout.
+const _alphabet = [
+	["C", "B#"],	//C
+	["C#", "Db"],	//C#
+	"D", 			//D
+	["D#", "Eb"],	//D#
+	"E", 			//E
+	["F", "E#"], 	//F
+	["F#", "Gb"],	//F#
+	"G", 			//G
+	["G#", "Ab"],	//G#
+	"A", 			//A
+	["A#", "Bb"],	//A#
+	["B", "Cb"]		//B
+]
+
+const _notes = {
+	0: "C",
+	1: "C#",
+	2: "D",
+	3: "D#",
+	4: "E",
+	5: "F",
+	6: "F#",
+	7: "G",
+	8: "G#",
+	9: "A",
+	10: "A#",
+	11: "B"
+}
+
 // interval library structure
 // also stores chords because chords are groupings of intervals
 // using freeze to prevent data from being altered
@@ -133,3 +165,5 @@ const _intervals = Object.freeze({
 		}]
 
 })
+
+//module.exports = global;
